Fix predicate encoding for prefix keys
This substantially reworks the way that pushdown of key range predicates
works in order to fix several bugs:

1) The old code was using EncodedKeyBuilder::BuildSuccessorKey(), which
didn't work correctly in a number of cases. In particular:
  - for string columns, it would increment the last character (e.g. "foo" ->
    "fop") which ignores the fact that there are many keys that could fall
    between the two (e.g. "foobar").
  - it only incremented the "encoded key", and not the "raw key", which meant
    that pushed predicates might work incorrectly on single-part keys.
  - for string keys, incrementing the raw encoded string could result in an
    invalid key (eg because the \x00\x01 escape sequence might mutate to \x00\x02,
    which is considered invalid).

   This is fixed by using the new row_key_util::Increment* functions to properly
   generate a new primary key tuple with the proper value. A number of new unit
   tests are added to verify this functionality.

   composite-pushdown-test is also modified to verify the behavior with string
   keys as parts of composites.

2) When handling a predicate on a prefix of a primary key, we previously generated
   encoded key ranges which were too short to be properly decoded back into 'tuple'
   format on the server side. This resulted in errors coming back from the tablet
   server when specifying such predicates on scans.

   The new code ensures that we always generate encoded keys which include all of
   the key "parts", thus ensuring that we can always decode them on the server
   side.

One difficulty that arose from change #2 above is in the handling of cases such as:

  primary key: (int8 a, string b)
  user predicate: a <= 100

In trying to convert this to a primary key range, we need to produce:

  primary key < (101, "")

Unfortunately, with in the current model, we always have _inclusive_ key ranges.
There is no inclusive key range which can capture the above predicate: it would
have to look like primary_key <= (100, MAX_STRING), but there is no such thing
as a MAX_STRING.

So, to fix this, I had to convert encoded key ranges to be exclusive rather than
inclusive throughout.

This actually has some nice side effects in terms of usability: for example, to
align a scanner to scan exactly one tablet, we can now directly pass the tablet
key range into the scanner's encoded key range, since the tablet's upper bound
is (and has always been) exclusive.

I had to modify the client sample.cc to reflect the bug fix above -- it was
previously depending on incorrect results when applying a range predicate to
the primary key.

Change-Id: I4f7096991607633a9ff442194a2b7454c5c63d1d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6851
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>