Tablet Servers should report FQDN to the master
Previously, the TS was just using GetHostname() to report its name
to the master. In one of our internal clusters, this was resulting in
a short non-qualified hostname that was not resolvable from other nodes.

This patch uses the FQDN. The FQDN is determined by the same method that
'hostname -f' uses.

Testing this is tricky since it's environmental. The new unit test just
calls the method and asserts that it returns OK. I manually tested on
the cluster in question and verified that a full FQDN was logged by
the test.

Thanks to Ricky Saltzer for the bug report.

Change-Id: I96a545975f68e81fc0537b38be1def7c144ffe88
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6859
Reviewed-by: Richard Saltzer <ricky@cloudera.com>
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins