KUDU-807: Retries are done in a tight loop when we can't resolve
This patch makes it so that when all the hosts are bad for a tablet or a master, we error out
right away.

Before this patch, a master specified with an unreachable hostname would retry forever with sleeps,
and a tablet with unreachable replicas would just tight loop sending RPCs to the master.

This patch also adds a test case. The testing for the master case is easy, but for the tablet case
we go through the back door and don't actually test the whole code path, since it would require
having the master send us bad locations. I did test it manually.

Change-Id: Id4119c0b257e3d6cd9f76dcb4e523ba5dd942f59
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6870
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>