KUDU-668: combine WritablePBContainerFile::Init into one append
The code makes two calls to Append(), increasing the chance of a corrupt
container file (i.e. killing the writer in between). If we combine the
preparation steps and make just one Append(), it will be more robust.

Note that the containers are still susceptible to truncation; that part of
KUDU-668 has not been addressed. Moreover, this only helps non-memory mapped
container files.

Change-Id: Ib8d1d21ca08d8b3e23c6d77b9abb27b922e01901
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6869
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins