debug-util: fix stack trace hex address generation
This fixes the dumping of hex stack traces to subtract 1 from each
instruction pointer. This fixes the results of 'addr2line' to
actually point to the _calls_ rather than the _return addresses_
of each function call on the stack.

I noticed this when trying to run 'addr2line' on some stack traces
seen for lock contention in a bolt-80 log. The stack was garbage until
I manually subtracted one from each address, at which point the stack
made a lot of sense.

It's difficult to write a reliable test for this, since any test would
be sensitive to the exact code generation of the compiler in use.
So, we'll just have to go based on the fact that the "subtract 1" trick
worked on a real stack trace.

Change-Id: I9f5145b5317ae376ffcfbad3370dd43dfe571747
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6894
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: Adar Dembo <adar@cloudera.com>