[java client] Don't delete tables that we might have to write to in BaseKuduTest
Every now and then, a Java unit test would fail and then it would hang for 50s (default sleep)
while shutting down the client. I finally found why: operations that need to be flushed in order
to close the client will never succeed since we delete the tables just before that.

The fact that operations going against deleted tables hang will be addressed in a different patch.

The deleting of tables was added because it's possible to reuse clusters when testing, but we
usually use unique table names so it's not a must have.

Change-Id: I351b7f3652ce69ca5ac476f7b09f93115de7d4e3
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6907
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins