build: download a snapshot build of clang 3.7 for sanitizer builds
This changes the thirdparty scripts to download a snapshot binary build
of clang 3.7, built from the Impala toolchain repository. The binary
is built on el6, but I verified that it also works on my newer Ubuntu
box.

The main reason for this change is that there are a few bugs in clang
3.5 which we've been hitting -- in particular, LSAN appears to deadlock
relatively frequently. This bug is fixed in clang trunk.

The revision chosen here is based on a snapshot that was used for a while
in building the Chromium project. I didn't use the most recent snapshot
currently used by Chromium because it crashes when trying to build Kudu.
This snapshot was used for a couple months by them and appears to build
OK.

I had to change rpc-test-base to use Kudu threads instead of boost::thread
because the TSAN in 3.7 somehow wasn't picking up the dependency between
the the new thread and the code that started it. Using Kudu thread fixed
the issue.

I additionally had to add a couple more suppressions for known issues,
due to the new TSAN being slightly more sensitive.

Change-Id: I2d9f64d431dd0e47fa1d581423b0b7676faefeda
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6917
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>