KUDU-664 (part 1): add version information to the build
This adds the basic integration of version numbers and build stamping.
In order to get this to work, I had to update cmake to 3.2.3, to
take advantage of a new feature which addresses the following bug:
  http://www.cmake.org/Bug/view.php?id=14963

This allows the build stamping to only re-run when important parameters
change (e.g the git hash or dirty state). If only the time changes, the
build information is not updated, which preserves fast builds during
normal development.

This hooks up the build information to the "-version" flag. We'll need
to add it in other places such as the web UIs, RPC info, etc.

Change-Id: I8994a2bb3c7461202630cc30a1aafced741ed6c4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6918
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>