fs: introduce path sets
This commit introduces the concept of a "path set": a group of related
filesystem roots and their relationships to other members of the group. Or,
in more concrete terms:
1. A UUID (globally unique) and an index (locally unique) for each path.
2. A set of the UUIDs belonging to all of the paths in the path set.

By persisting a path set in each path instance, we can safely refer to each
root path by its index instead of by its UUID. This in turn will allow the
file block manager to use 64-bit block IDs while still embedding enough
information in each ID to map to a specific root path. That's the underlying
motivation behind all of this, though path sets may have other uses in the
future, such as enabling a safe add/remove disk workflow.

Change-Id: I63f9a2656dce7707cb273d0fc73e2e38b8504c26
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6934
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>