fs: 64-bit block IDs
The use of std::string to store block IDs was convenient yet subpar:
1. Strings make BlockIdPB more expensive to serialize/deserialize.
2. They consume more space on disk and in memory.
3. They make accurate memory tracking difficult because they contain private
   heap pointers.

Could we use 64-bits instead? That would provide a large enough block ID
space to satisfy our scale targets while staying both cheap and compact. The
log block manager is easily adapted to 64-bit block IDs as it only uses them
as completely opaque keys into its global block map.

The file block manager is more complicated; it uses long (256-bit) IDs that
combine the root path and the block's UUIDs. The latter can be easily
replaced by a 64-bit random integer, but what of the former? Using the path
set abstraction from the previous commit, we can safely use a path's index
in lieu of its UUID. If we reserve two bytes for this index and squish the
aforementioned 64-bit random integer into the remaining six bytes, we'll
have a working solution.

It's a little clunky and there will be many block collisions if pushed to
the limit, but that's probably good enough for the file block manager since
it's intended to be a fallback for the log block manager at this point.

Change-Id: I063318b7631093f242f7d996745270b4bca54e63
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/6935
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>