KUDU-382 (1/many): allow RowChangeList to be encoded/decoded without a schema
This implements one of the changes described by the Alter Schema 2.0 design
document [1]. The encoding of UPDATE RowChangeLists is changed such that they
can be decoded with no knowledge of the schema.

This allows DeltaFiles, DeltaMemStores, and mutations in MRS to no longer be
associated with a schema. Instead, we can use the scanner's projection schema,
and simply skip over any updated columns which are not part of the projection.
In the case of delta flush or compaction, we can blindly flush deltas to disk
without knowledge of the schema at all.

As described in the doc, this will enable two benefits:
1) it simplifies the handling of AlterSchema, since we no longer need to flush
   DMS when altering a table.
2) it will reduce memory usage significantly, since we no longer will need to
   keep Schema copies in DMS/DeltaFileReader.

This patch is mostly prep work within RowChangeList and the various iterators.
The next patch in this series removes the actual Schema members.

[1] https://docs.google.com/document/d/1n0OOctNvTv33XP0hGQox0tP5kuVYsUs-QniFy_UU2PI/edit#

Change-Id: I4b963e37637e02062c68d7414a3de6041ec6b9da
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7001
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>