KUDU-382 (2/many): remove Schema members from DeltaFileReader and DeltaMemStore
Now that RowChangeLists can be decoded and encoded without a Schema object, we
no longer need to keep a Schema around in DMS and DFR. When we alter the
schema, we no longer need to propagate the alter to the delta tracker at all.

This change removes the Schema members from those classes, and also removes the
serialized Schema from the DeltaFile on-disk format. It also changes the
serialized DeltaStatsPB to record the stats keyed by column ID instead of
column index.

This change should reduce memory consumption substantially: we remove at least
two copies of the schema for every DiskRowSet (one for the DMS, one for the
UNDO delta file, and potentially more if there have been any updates).

The DeltaTracker still maintains a Schema member. This will be removed in a
subsequent patch. Additionally, the DiskRowSet itself can also be decoupled
from the Schema. Again, this will be a later patch in this series.

Change-Id: I5b8dafa83af1acbd2b72f435f83c49f03e1a45b9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7002
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins