KUDU-382 (3/many): rename FilterColumnsAndCollectDeltas
Some of the definitions of this function called the parameter column_indexes,
even though in fact it was taking a list of column IDs. This patch simply
renames the function and clarifies the parameter name in all usage sites.

An earlier patch already fixed the usage to pass column IDs.

Change-Id: Ibb64db0e7613ef3c36f2c35263d6c0400043648d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7005
Reviewed-by: Adar Dembo <adar@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins