KUDU-382 (4/many): remove schema_ member from DeltaTracker, remove RowSet::AlterSchema
This removes the schema_ from DeltaTracker. This results in
DiskRowSet::AlterSchema becoming a no-op, so I was able to also remove it and
its callers.

This made one of the tests in mt-rowset_delta_compaction-test essentially a
no-op as well, so I removed it and added a TODO in its place to get better test
coverage of updates with a changing schema.

Note that this does not yet remove the per-DRS schema stored within
RowSetMetadata.  That will be addressed in a later patch -- first, we need to
make CFileSet keep its columns by ID rather than index.

Change-Id: Icf31b361636477dba323e77ec93d0b79caba00d2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7006
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: David Alves <david.alves@cloudera.com>