util: memory reporting APIs for various utility objects
Inspired by the Firefox memory reporting APIs [1][2], this patch introduces
a new set of APIs with which to measure the memory usage of an object. I'm
deviating from the Firefox design in a few ways:
1. We call malloc_usable_size() directly instead of passing it around as a
   function pointer, an indirection that doesn't seem necessary for Kudu.
2. The memory_footprint_excluding_this() variant is omitted if the class must be
   allocated on the heap (i.e. is refcounted).
3. Both variants are omitted if usage is trivially measured (i.e. is always
   either heap or stack allocated, and contains no embedded heap pointers).

In this patch I've rolled out the APIs to classes are embedded in other
"high value" classes (i.e. those whose instance count scales with the amount
of data under management).

1. https://blog.mozilla.org/nnethercote/2015/06/03/measuring-data-structure-sizes-firefox-c-vs-servo-rust
2. https://developer.mozilla.org/en-US/docs/Mozilla/Performance/Memory_reporting

Change-Id: I2e252505741067192920d4397536fbc989ca4d1f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7025
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>