fs: memory reporting APIs for block managers and their dependents
In this patch we roll out the APIs to the block managers and their
dependents. Of note: ReadableBlocks no longer consume/release memory; in a
subsequent patch this will become the responsibility of CFileReaders. I
think this makes sense because it associates every ReadableBlock with the
tablet that opened it.

Change-Id: I9ccbdbdf97f60c95071c206a740e8997176bd7c1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7026
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>