cfile: memory tracking for cfile and bloom readers
Besides rolling out the new memory footprint APIs, this patch provides a
taste of how object memory tracking will work:
1. The tracked object embeds a ScopedTrackedConsumption (henceforth STC).
2. At construction time, the STC is initialized with memory_footprint().
   This requires that the object is allocated on the heap.
3. If the object's footprint changes, the STC is reset with the new value.
4. When the destructor is called, the STC goes out of scope and releases
   the tracked memory.

Both newly tracked objects exhibit #3 when Init() is called. Furthermore,
BloomFileReader has an interesting variation on memory_footprint(); the
embedded CFileReader is excluded from measurement because its footprint
is tracked separately. This is a theme that will be repeated in other
objects in the future.

Change-Id: Ib6216eac5f4e0e3fd87e6cc3edb7e9c1a0e313d6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7028
Tested-by: jenkins
Reviewed-by: Todd Lipcon <todd@cloudera.com>