KUDU-382 (6/many): remove schema() from RowSet metadata, interface
This removes the schema member from RowSetMetadata and the corresponding
column schema information from its protobuf form. Instead, we track the
schema only in the TabletMetadata, and CFileSet looks to the TabletMetadata
instance to find the key column ID.

After this change, it was possible to remove schema() from the RowSet
interface entirely.

In order to allow lockless access to the tablet-level Schema, and also
not have to make Schema ref-counted, the TabletMetadata instance currently
maintains all previous versions of the Schema in a vector. Since we don't
anticipate thousands of alter-table operations on a given table, this is likely
a reasonable short-cut.

Change-Id: Ibad2ed8bd4c1f0a1514de9f290b62d7b85890127
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7047
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins