KUDU-461. Don't re-use column IDs when altering a table
This adds a field to the master metadata for a table which specifies
the next column ID that should be assigned.

If this is set, then new columns are assigned starting with that ID.
If it is not set, then we use one higher than the highest currently-used
ID.

This re-enables AlterTableTest.TestDropAndAddNewColumn.

Change-Id: I402d70ef4b4eb8fde3f3e101aa9fab4256d0970c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7049
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>