build: merge exported client library build into regular build
This patch rewrites the exported Kudu client library build logic such that
it's part of the regular build. Now you don't have to remember whether the
libkudu_client.so in your build tree is the exported client or not, and the
Python bindings can be used all the time.

To build the exported client, we need to compile parts of Kudu with
different flags. Unfortunately, there's no way to reuse these objects in the
regular build as symbol visibility is all messed up, so we're forced to eat
the added compilation time. The full list is gutil, util, common, krpc,
client, and various proto targets. They are built twice using the new
ADD_LIBRARY_TWICE() function.

The exported client is a "fat" shared object that statically includes all of
the client's dependencies (direct and transitive). This presents a challenge
for KUDU_LINK=d builds: how do we provide the exported client with static
archives of both Kudu and thirdparty targets?
- For Kudu libraries, the ADD_LIBRARY_TWICE() wrapper ensures that the
  "exported" variant is built as a static archive.
- For thirdparty libraries, ADD_THIRDPARTY_LIB() now always publishes a
  static variant of each target.

Why must we be so exhaustive? For example, in KUDU_LINK=s builds, why must
we create new thirdparty targets instead of reusing the existing (static)
ones? The answer is two-fold:
1. In our CMakeLists.txt, we intermingle thirdparty and Kudu targets when
   calling target_link_libraries().
2. We have circular dependencies amongst our subdirectories. For example,
   consensus_proto (consensus) -> tserver_admin_proto (tserver) ->
   tserver_proto (tserver) -> consensus_metadata_proto (consensus).
Taken together, this breaks "if foo is a valid target then depend on foo,
otherwise depend on foo_exported" conditions.

On the Jenkins front, this patch makes the CLIENT build type a regular
build. A side effect is that we're now directly exposed to KUDU-455, because
we're still using the system gcc on our el6 boxes. After the patch is merged
I'll remove the CLIENT entry from kudu-gerrit.

Change-Id: I673b9ccc7ee0a476bf55471d53d036635d93e4d6
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7069
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins