Plumb RaftPeerPB to consensus queue
This takes advantage of the prior patch in this series to look up the bound RPC
address of the tablet server, create a local RaftPeerPB, and pass it into
RaftConsensus. This will allow the leader to pass its own peer address to a new
node in order to trigger a remote bootstrap.

Change-Id: Ia81dd05ab977b2369c534e1c8e31a2b225b0acc4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7113
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins