KUDU-621 (part 1): snapshot the state of flushed stores at the beginning of bootstrap
Currently, bootstrap uses the TabletMetadata object to determine whether operations
in the log need to be replayed. However, if we were to flush or compact during
replay, the TabletMetadata object would change state and we could make
incorrect decisions about what needs to be replayed.

This adds a simple class in the bootstrap process which snapshots the relevant
information from TabletMetadata, so that we could safely flush/compact during
bootstrap.

This does not actually add flushing/compaction during bootstrap: doing so is a
little bit tricky, since the maintenance manager operations are registered on a
tablet peer, and the tablet peer is not yet fully initialized during bootstrap.
However, this _is_ enough to safely call Flush() in the context of an AlterTable()
call and not break durability. This is important to fix KUDU-181 and KUDU-462.

Change-Id: I5f1af7d9f565ecfe411d8140409043e1cda8b873
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7066
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>