KUDU-382 (9/many): fix compaction if a delta applies to a deleted column
This adds a new test which exercises the following sequence:
- an update is applied to a row which is in a DRS
- one of the columns that was updated is dropped
- we run a compaction

This was previously crashing with a CHECK failure, since we incorrectly
asserted that we should not get "column not found" when trying to apply
deltas during a compaction. In fact, this is fine, and we should just
drop the delta corresponding to the deleted column.

A happy side-effect of this change is that it should improve compaction
performance: a profile of a compaction-heavy workload shows
ColumnSchema::Equals being called from ApplyMutationsAndGenerateUndos as the
fifth highest CPU consumer. This patch removes that call.

Change-Id: I0ff0faa2b436304494e0b4bf805fe3579370bc1a
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7073
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>