KUDU-382, KUDU-701, KUDU-363: randomized integration test for alter table
This adds an integration test which performs a randomized sequence of operations
against a tablet server, including inserts, updates, deletes, and table alterations.

This fairly reliably reproduced all of the above bugs as well as a few others which
have recently been fixed under the umbrella of KUDU-382.

The test uses a new flag in PosixEnv to disable fsync server-wide. This improves the
speed of the test substantially, since the test does a lot of small flushes and
metadata changes.

The test also has the ability to restart the tablet server randomly during the
sequence. This portion is commented out while waiting for KUDU-181/KUDU-462
to be resolved.

Change-Id: I754e4d7d9aff7b407b9a888f4d6676c63846f4f4
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7076
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins