id_mapping: fix a bug when growing hashtable
This fixes a CHECK failure that would be seen with some sequences of
insertions into the map. The issue was that we were assuming that,
if we doubled the size of the slot vector, we would have zero collisions
in the newly doubled map. This isn't the case, since we do linear probing.

Change-Id: If8d9640d28f5936dff7f2cc0e3565886ce130cce
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7075
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>