cfile_reader: retain reference to dictionary block in iterator
This fixes a crash seen while testing dictionary coding on a real workload.
The CFileIterator was initializing a block decoder from the dictionary block,
but then releasing the block cache handle. So, if the block cache evicted the
dictionary block while the iterator was still alive, then we'd read invalid
heap pointers.

I tried to write a unit test, but couldn't come up with anything simple.
Instead, I added a new sanity check assertion which would fail on the
cluster with the bug. Additionally, cluster testing the time series workload
failed without this patch after a few hours, but has run for about half a day
after the patch.

Change-Id: Ic3953c777fe3d35a4957d7a238b6d00a59fc1043
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7086
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>