KUDU-848 (part 1): improve client API for specifying range predicates, add KuduValue
As noted in KUDU-848, the client API up until now relied on the user
constructing various void* pointers, which we then didn't even aggressively
copy. This made it very difficult to use the API (even confusing Kudu
developers at times into thinking we had bugs in predicate evaluation!)

Additionally, the old predicate API always took in upper and lower bounds
together, rather than a more standard-looking "comparison operator" based API.
The user was expected to pass 'NULL' to signify that no bound was to be used.
But, that will be confusing when we also add the ability to match equality on
NULL values.

This patch changes the API to use a new KuduValue helper class which copies the
user's value into a sort of tagged type. The predicate API itself is moved to
KuduTable::NewComparisonPredicate(...) which looks a lot more like a typical
'foo <= 12345' type predicate a user would see in SQL.

Under the hood, we still convert the user's predicates into a
ColumnRangePredicate, but this new API will give us more flexibility to evolve
in the future without breakage, and no longer relies on them keeping values
live for the duration of the scanner.

A future patch will also use KuduValue to specify constants such as the default
value for a column when creating a table.

Change-Id: I0028d57b3d793f5bf3d4c9346f4ca64c7262703f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7089
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Todd Lipcon <todd@cloudera.com>
Reviewed-by: David Alves <david.alves@cloudera.com>