KUDU-701. Fix potential data loss during concurrent tablet metadata operations
This changes the metadata flush path to keep a local copy of orphaned_blocks
corresponding to the superblock version being written. The flushing thread only
tries to delete those blocks from disk. This prevents the race described in
KUDU-701 where another thread may add to the orphaned_blocks_ list while the
first thread is in the process of flushing the superblock.

This patch doesn't add a new regression test, but the regression test for
KUDU-382 (alter_table-randomized-test) reproduces this bug fairly reliably
(~3% of the time). With this patch, it doesn't reproduce.

Change-Id: Ifb386efd934a07697781e7146bd2e8e2ab22a2cf
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7093
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins