Remove the "milestone 3" TPCH demo
This code was originally written for a demo a couple of years ago, and hasn't
been actively maintained or used since then. Rather than spending more time
maintaining it as the client API changes, I think it's better to remove it.

Nowadays, we have better ways to demo Kudu.

If for some reason we'd like to revive this as a demo, we can recover the code
from the git history.

Change-Id: I4dd0f27d3cba9acf2a39d0184b4d9be0a3059878
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7109
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins