KUDU-664 (part 2): expose version info in webserver and RPC
This adds the version info to the main web page for each service, and also adds
a new GetStatus RPC to the base service. The new RPC exposes the existing
instance information as well as the version info.

We don't currently have automated tests for these management RPCs, but I manually
verified that it worked against a local master.

Change-Id: I1510bc76ce115fb5d9f957fc540a8ac5efd3a4fe
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7146
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>