KUDU-181, KUDU-462. Bootstrap should replay operations with a "point-in-time" schema
This changes the bootstrap process to rewind the tablet schema to the schema
found in the header of the first log segment before beginning replay. Then, as
each new AlterSchema operation is replayed, the tablet moves forward through
the schema history.

To make this a bit more safe, I changed the 'open_' boolean in Tablet to a
State enum, and ensured that the "rewind schema" function is only used in the
bootstrap scenario.

There are a couple more pending issues around the handling of Alter Schema
operations in the log which need attention, but this is enough to pass the
existing tests as well as a few new ones. The issues are noted in TODOs and
will be addressed as follow-ups.

Change-Id: Ic05698172550c0fe8865d81df65b20d47789af28
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7068
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>