KUDU-664 (part 3): allow git hash to be provided on the cmake command line
The Kudu packaging build is run outside of a git repository on each build
slave. Thus, we need a mechanism to manually feed the git hash into
gen_version_info.py. This patch adds that mechanism. A separate patch (to
cdh-package.git) makes use of it in the packaging build.

I tested this patch manually by running cmake both with and without
-DKUDU_GIT_HASH and verifying the contents of version_defines.h.

Change-Id: Iab19c016648c5ffbd67c04b4fe18212bceae51a1
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7163
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins