cfile: reduce number of transitive includes of encoding classes
I noticed that type_encodings.h was previously including all of the
encoder/decoder header files, many of which are heavily templated.
type_encodings.h is encoded by a number of other header files, so we ended up
having to do a lot of compilation work whenever any encoding changed, and a lot
of extra CPU on expanding and compiling these templates.

This simply moves the heavy work of type_encodings.h into the .cc file,
and also reduces a few other unnecessary header includes where forward
declarations are sufficient.

Change-Id: I004562260c2015603d8f8f14a013b5f58148c48c
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7165
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins
Reviewed-by: Guangxiang Du <guangxiang.du@cloudera.com>