monitoring: add a /dump-entities page on the master
This web page dumps a JSON-formatted list of all of the tablets and tables in
the system catalog, along with basic attribute information and replica
locations. This will be used by CM's service monitoring daemon to construct its
internal database of time series entities.

I also added two new attributes to the MDL which are exposed via this web page.

Change-Id: I1aa04222e4053cb06d7e6c04a3d1d038c7848807
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7175
Tested-by: jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>