csd: fix build and add validation
Commit 4ce99f0 forgot to define the new metric attributes; this patch
includes that.

Additionally, I've modified the CSD build process to validate the SDL and
MDL. We've got to validate against a particular CSD schema version; for the
time being, I've chosen the "latest" schema (i.e. whatever will ship in
5.5.0). If/when 5.5.0 breaks backwards compatibility with what shipped in
5.4.3, we'll need to fork the CSD anyway, so forking the validation won't be
an issue.

Note that the current 5.5.0-SNAPSHOT validator doesn't validate MDL files,
but I've got a patch outstanding to fix that (in the CM repo). When applied
without the generate_mdl.py patch, it causes the Kudu CSD build to fail.

Change-Id: I47d914984bfcdaa581e632fe2fd7419dcd9a2f23
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7190
Reviewed-by: Christian Leroy <cleroy@cloudera.com>
Tested-by: jenkins