KUDU-755, KUDU-849. Tablet locations should not disappear while BOOTSTRAPPING
This fixes a bug where, when a TS started up and reported tablets that were
bootstrapping, the master would remove those replicas from the list of tablet
locations. The reason was that, on a full tablet report, we removed all
replicas from the TS before adding back the ones that were reported. However,
we used the committed consensus state to add back the reported replicas, and a
bootstrapping tablet does not report its consensus state. Thus, while a tablet
was bootstrapping, it wouldn't show up in the locations list.

This also fixes KUDU-849, a more rare race in which an Alter command could
be missed if it arrived exactly as a tablet finished bootstrapping. The race
was:
1) TS does a full report, which removes the tablet location from the list of
   replicas (the bug decscribed above)
2) Alter Table command arrives at the master. It iterates over locations, but
   misses this location because it has been removed. The schema version has
   not yet been updated.
3) The TS does an incremental report, which adds back the replica. Because the
   schema version was not yet bumped, it doesn't see that an ALTER is needed.
4) The Alter Table command continues processing, and changes the schema to
   the new version.

Hence, the alter would never get sent to the tablet replica.

Before the patch, I was able to reproduce this reliably by inserting a sleep
in HandleReportedTablet() after it checks the schema version but before it
adds the replica. With the patch, it doesn't reproduce.

This also could have been fixed by changing the order of HandleReportedTablet
to add the tablet location before checking the schema version. However, I think
these code paths need a more general overhaul in order to support altering a
replicated tablet, so I didn't want to do any major restructuring of the code
that's about to change. Additionally, this commit has the side effect of fixing
the other existing bug KUDU-755.

Change-Id: If9e51fda0dda17b6c98e98ea324d1f24e18099f2
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7210
Reviewed-by: Jean-Daniel Cryans <jdcryans@cloudera.com>
Tested-by: jenkins