KUDU-304 (part 2): refactor CatalogManager hierarchy for background tasks
This refactors the code in the CatalogManager which is responsible for
repeatedly retrying an RPC against a tablet server. These retryable RPCs
currently are always targeted at a specific _replica_ (i.e a specific tablet ID
on a specific tablet server). To properly support alter-table, though, we need
to repeatedly send the RPC only to the _leader_ replica, and retry on the new
leader if the leader changes.

This patch extracts the code which is responsible for picking the target tablet
server into a simple TSPicker interface. The one implementation of this
interface maintains the existing behavior of sending the RPCs to a specific
tablet server, identified by UUID. The next patch in this series will add a new
implementation which picks the leader.

Change-Id: I1fa6f93087eb51b12c2271b52737d17b9b9e7a8d
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7273
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: jenkins