Only change TabletPeer state to RUNNING after consensus starts
This should fix a source of flakiness in alter_table-randomized-test
where we see the following assertion failure:

F0725 01:32:48.067261 12247 local_consensus.cc:100] Check failed: state_ >= kConfiguring (1 vs. 2)

This race occurred if an RPC arrived for a tablet after the peer state switched
to RUNNING, but before Consensus was ready to service requests.

Change-Id: I2a3d129c02c2ceba44addf017d04db905389985f
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7231
Reviewed-by: David Alves <david.alves@cloudera.com>
Tested-by: jenkins