ksck_remote-test: fix flakiness of TestChecksumTimeout
This test case was generating a single batch with 10000 rows in it,
which was large enough to often time out in TSAN builds. There's no
reason that the test needs such a large batch, so this patch changes it
to flush smaller batches of 1000 rows each.

Without the patch, this failed 22/100 TSAN test runs. With the patch,
it passed 100%.

Change-Id: I16f49293b8d8dc71ebe614652b32e52b1f763b36
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7232
Reviewed-by: Michael Percy <mpercy@cloudera.com>
Tested-by: jenkins