mem_tracker: soft memory limit
This patch introduces the concept of a soft memory limit, which may be
exceeded even before consumption exceeds the regular memory limit.
Every time the soft limit is checked, whether or not it was exceeded depends
on the current consumption and a randomly generated number; the higher the
consumption the greater the chance. This is useful for backpressure
applications that smoothly increase/decrease rather than abruptly
enable/disable.

The new limit is expressed as a percentage of the regular limit, and for
simplicity, I've defined it as a gflag rather than allowing it to be
customized for each tracker. It does not check the limit on all ancestral
trackers to be consistent with LimitExceeded(), which in practice means it's
only really useful on the root tracker.

I also renamed and rescaled the hard memory limit to be more user friendly.

Change-Id: Ifdcc6af7bd62a809dfbffbee63908a454cb9a1b9
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/7254
Reviewed-by: Todd Lipcon <todd@cloudera.com>
Tested-by: jenkins